requirements.txt
imageio>=2.33.0
numpy>=1.24.0
Pillow>=10.0.0
